cleanup export area, add responseId to the export, prepare importing data
